Move rest of stuff out of main to this generate package for now
I can admit.. naming is not one of my strengths.
